Refactor: Split ShapeGamepadAxis into ShapeGamepadJoyAxis and ShapeGamepadTriggerAxis. Refactor: Split InputTypeGamepadAxis into InputTypeGamepadJoyAxis and InputTypeGamepadTriggerAxis. Feat: Added new static create methods to InputTypeGamepadJoyAxis and InputTypeGamepadTriggerAxis for each joy/trigger.
Fixed all errors in files that used the old ShapeGamepadAxis enum or the old InputTypeGamepadAxis class.

The input system is setup to only use a single deadzone value. Mouse alrady had its different axis types split into multiple enums but the GamepadDevice had joy and trigger axis in one enum. To simplify and unify the system gamepad was now split into 2 seperate axis type as well to make deadzones work better and to be the same as the MouseDevice.